+++ docs/release-notes.md
@@ -3,6 +3,7 @@
 * Fixed issue with empty files opening the configuration view.
 * Fixed issues with opening large resource/archive files.
 * Fixed failures with creating self-extracting archives from large archive files.
+* Fixed issues updating Siege Studio if the version falls in the next year.
 
 # Siege Studio + Siege Launcher 26.15 Development Release
 * Added initial interface to configure controller types if the controller is not detected correctly.